requirements.txt
Django==4.1.4
djangorestframework==3.14.0
dj-database-url==1.2.0
gunicorn==20.1.0
Pillow==9.3.0
psycopg2-binary==2.9.5
stripe==5.0.0
uvicorn==0.20.0
whitenoise==6.2.0
